# swagger/portal-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: wicked.haufe.io Portal API
  description: The API which the API Portal uses to store and retrieve information
 
# Under securityDefinitions we declare which security schemes can be used.
# These definitions do not affect to the operations, but can be applied
# using the 'security' key at the global level or the operation level.
# In this sample, we'll see how to apply the security definition globally,
# and override it at the operation level.
#
# Note that for Access Code Flow, you must provide the Token URL.
securityDefinitions:
  # For personalized access to data
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/auth/local/portal-api/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/auth/local/portal-api/oauth2/authorize
    flow: accessCode
    scopes:
      read_plans: Read plans
      read_groups: Read user group definitions
      read_content: Read custom/markdown/HTML content
      health: Read out health information
      read_templates: Read templates
      read_auth_servers: Read authorization servers
      read_apis: Read API definitions
      read_users: Read user information
      write_users: Write user information
      read_applications: Read application data
      write_applications: Write application data
      read_subscriptions: Read subscriptions
      write_subscriptions: Write/edit/delete subscriptions
      read_approvals: Read pending approvals
      read_verifications: Read email and password verifications
      write_verifications: Create or delete verifications
      read_registrations: Read user registrations
      write_registrations: Write user registrations
      read_grants: Read user API grants
      write_grants: Write user API grants
      webhooks: Access and change webhooks
  
  implicit:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/portal-api/v1/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/portal-auth/api/portal-api
    flow: accessCode
    scopes:
      read_plans: Read plans
...
produces:
  - application/hal+json
host: 'localhost:3001'
basePath: /portal-api/v1
...
paths:
...
  '/verifications':
    post:
      security:
      - clientCredentials: []
      tags:
        - verifications
      description: >-
        Registers for a new verification of email or a request for password
        reset
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/verificationType'
      responses:
        '204':
          description: >-
            No content. The API will not return whether it was successful or
            not.
        '400':
          description: Bad request
    get:
      security:
      - accessCode: ['read_verifications']
      - implicit: ['read_verifications']
      - password: ['read_verifications']
      tags:
        - verifications
      description: Returns a list of pending verifications (admin)
      responses:
        '200':
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/verificationType'
        '403':
          description: Not allowed.
  '/verifications/{verificationId}':
    get:
      security:
      - clientCredentials: []
      - accessCode: []
      - implicit: []
      - password: []
      tags:
        - verifications
      parameters:
        - name: verificationId
          in: path
          type: string
          required: true
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/verificationType'
        '404':
          description: Not found
    delete:
      security:
      - clientCredentials: []
      - accessCode: []
      - implicit: []
      - password: []
      tags:
        - verifications
      parameters:
        - name: verificationId
          in: path
          type: string
          required: true
      responses:
        '204':
          description: 'Success, no content'
        '404':
          description: Not found.
...
definitions:
...
  verificationType:
    type: object
    properties:
      id:
        description: Verification ID
        type: string
      type:
        description: Type of verification
        type: string
        enum:
          - email
          - lostpassword
      email:
        type: string
        description: The E-Mail address concerned in the validation
      userId:
        type: string
        description: 'When retrieved, this field will contain the associated user ID'
...